Change DoPipelineBarrier to take a size_t count not uint32_t
* Since we're always passing an array which naturally has a size_t count, this
  makes more sense.

+++ renderdoc/driver/vulkan/vk_common.cpp
@@ -441,31 +441,31 @@ int StageIndex(VkShaderStageFlagBits stageFlag)
   return 0;
 }
 
-void DoPipelineBarrier(VkCommandBuffer cmd, uint32_t count, const VkImageMemoryBarrier *barriers)
+void DoPipelineBarrier(VkCommandBuffer cmd, size_t count, const VkImageMemoryBarrier *barriers)
 {
   RDCASSERT(cmd != VK_NULL_HANDLE);
   ObjDisp(cmd)->CmdPipelineBarrier(Unwrap(cmd), VK_PIPELINE_STAGE_ALL_COMMANDS_BIT,
                                    VK_PIPELINE_STAGE_ALL_COMMANDS_BIT, 0, 0,
-                                   NULL,                // global memory barriers
-                                   0, NULL,             // buffer memory barriers
-                                   count, barriers);    // image memory barriers
+                                   NULL,                          // global memory barriers
+                                   0, NULL,                       // buffer memory barriers
+                                   (uint32_t)count, barriers);    // image memory barriers
 }
 
-void DoPipelineBarrier(VkCommandBuffer cmd, uint32_t count, const VkBufferMemoryBarrier *barriers)
+void DoPipelineBarrier(VkCommandBuffer cmd, size_t count, const VkBufferMemoryBarrier *barriers)
 {
   RDCASSERT(cmd != VK_NULL_HANDLE);
   ObjDisp(cmd)->CmdPipelineBarrier(Unwrap(cmd), VK_PIPELINE_STAGE_ALL_COMMANDS_BIT,
                                    VK_PIPELINE_STAGE_ALL_COMMANDS_BIT, 0, 0,
-                                   NULL,               // global memory barriers
-                                   count, barriers,    // buffer memory barriers
-                                   0, NULL);           // image memory barriers
+                                   NULL,                         // global memory barriers
+                                   (uint32_t)count, barriers,    // buffer memory barriers
+                                   0, NULL);                     // image memory barriers
 }
 
-void DoPipelineBarrier(VkCommandBuffer cmd, uint32_t count, const VkMemoryBarrier *barriers)
+void DoPipelineBarrier(VkCommandBuffer cmd, size_t count, const VkMemoryBarrier *barriers)
 {
   RDCASSERT(cmd != VK_NULL_HANDLE);
   ObjDisp(cmd)->CmdPipelineBarrier(Unwrap(cmd), VK_PIPELINE_STAGE_ALL_COMMANDS_BIT,
-                                   VK_PIPELINE_STAGE_ALL_COMMANDS_BIT, 0, count,
+                                   VK_PIPELINE_STAGE_ALL_COMMANDS_BIT, 0, (uint32_t)count,
                                    barriers,    // global memory barriers
                                    0, NULL,     // buffer memory barriers
                                    0, NULL);    // image memory barriers
